+++ .github/workflows/main.yml
@@ -23,7 +23,7 @@ jobs:
           cp LICENSE package
           cp ModInfo.xml package
           cp README.md package
-          cp -r HighPingKicker/latest/* package
+          cp -r latest/* package
           cd package
           zip -qr "../${{ github.event.repository.name }}.zip" *
 
@@ -41,7 +41,7 @@ jobs:
           tag: ${{ env.version }}
           commit: main
           name: ${{ github.event.pull_request.title }}
-          body: "## Summary\n${{ github.event.pull_request.body }}"
+          body: "${{ github.event.pull_request.body }}"
           generateReleaseNotes: true
           artifacts: "${{ github.event.repository.name }}.zip"
           prerelease: ${{ env.prerelease }}
