+++ LagShield/ModApi.cs
@@ -1,8 +1,8 @@
-﻿namespace HighPingKicker {
+﻿namespace LagShield {
     public class ModApi : IModApi {
         public void InitMod(Mod _modInstance) {
             Service.Load();
-            ModEvents.SavePlayerData.RegisterHandler(Service.CheckPing);
+            ModEvents.SavePlayerData.RegisterHandler(Service.CheckLag);
         }
     }
 }
