+++ ModInfo.xml
@@ -1,9 +1,9 @@
 ﻿<xml>
     <ModInfo>
-        <Name value="High Ping Kicker" />
+        <Name value="Lag Shield" />
         <Description value="Automatically monitor and kick/ban players with excessively high ping" />
         <Author value="Jonathan Robertson (Kanaverum)" />
-        <Version value="1.0.1" />
-        <Website value="https://github.com/jonathan-robertson/high-ping-kicker" />
+        <Version value="1.0.0" />
+        <Website value="https://github.com/jonathan-robertson/lag-shield" />
     </ModInfo>
 </xml>
