+++ lag-shield.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32328.378
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HighPingKicker", "HighPingKicker\HighPingKicker.csproj", "{84DC5C4F-693B-4CD5-8EF3-4255508C2AF9}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{FA1A2DDC-660B-4FF4-8F1C-38A43229FA74}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
@@ -14,16 +12,18 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LagShield", "LagShield\LagShield.csproj", "{BF9A0785-D62C-4878-A483-AA2653C4DB12}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{84DC5C4F-693B-4CD5-8EF3-4255508C2AF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{84DC5C4F-693B-4CD5-8EF3-4255508C2AF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{84DC5C4F-693B-4CD5-8EF3-4255508C2AF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{84DC5C4F-693B-4CD5-8EF3-4255508C2AF9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BF9A0785-D62C-4878-A483-AA2653C4DB12}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BF9A0785-D62C-4878-A483-AA2653C4DB12}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BF9A0785-D62C-4878-A483-AA2653C4DB12}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BF9A0785-D62C-4878-A483-AA2653C4DB12}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
